+++ tests/Unit/DSC_PfxImport.Tests.ps1
@@ -1,5 +1,8 @@
-$script:dscModuleName = 'ComputerManagementDsc'
-$script:dscResourceName = 'DSC_CertificateImport'
+[System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingConvertToSecureStringWithPlainText', '')]
+param ()
+
+$script:dscModuleName = 'CertificateDsc'
+$script:dscResourceName = 'DSC_PfxImport'
 
 function Invoke-TestSetup
 {
@@ -44,23 +47,41 @@ try
             }
         ) -join ''
 
-        $testFile = 'test.cer'
-        $certificateFriendlyName = 'Test Certificate Friendly Name'
+        $testFile = 'test.pfx'
+
+        $testUsername = 'DummyUsername'
+        $testPassword = 'DummyPassword'
+        $testCredential = New-Object `
+            -TypeName System.Management.Automation.PSCredential `
+            -ArgumentList $testUsername, (ConvertTo-SecureString $testPassword -AsPlainText -Force)
 
         $validPath = "TestDrive:\$testFile"
         $validCertPath = "Cert:\LocalMachine\My"
+        $validCertFullPath = '{0}\{1}' -f $validCertPath, $validThumbprint
 
-        $validCertificate_mock = {
+        $certificateFriendlyName = 'Test Certificate Friendly Name'
+
+        $validCertificateWithPrivateKey_mock = {
             @{
-                Thumbprint   = $validThumbprint
-                FriendlyName = $certificateFriendlyName
+                Thumbprint    = $validThumbprint
+                HasPrivateKey = $true
+                FriendlyName  = $certificateFriendlyName
+            }
+        }
+
+        $validCertificateWithoutPrivateKey_mock = {
+            @{
+                Thumbprint    = $validThumbprint
+                HasPrivateKey = $false
+                FriendlyName  = $certificateFriendlyName
             }
         }
 
         $validCertificateWithDifferentFriendlyName_mock = {
             @{
-                Thumbprint   = $validThumbprint
-                FriendlyName = 'Different Friendly Name'
+                Thumbprint    = $validThumbprint
+                HasPrivateKey = $true
+                FriendlyName  = 'Different Friendly Name'
             }
         }
 
@@ -74,9 +95,11 @@ try
                 $Store -eq 'My'
         }
 
-        $importCertificateEx_parameterfilter = {
+        $importPfxCertificate_parameterfilter = {
             $CertStoreLocation -eq $validCertPath -and `
-                $FilePath -eq $validPath
+                $FilePath -eq $validPath -and `
+                $Exportable -eq $True -and `
+                $Password -eq $testCredential.Password
         }
 
         $setCertificateFriendlyNameInCertificateStore_parameterfilter = {
@@ -93,12 +116,14 @@ try
         }
 
         $presentParams = @{
-            Thumbprint   = $validThumbprint
-            Path         = $validPath
-            Ensure       = 'Present'
-            Location     = 'LocalMachine'
-            Store        = 'My'
-            Verbose      = $true
+            Thumbprint = $validThumbprint
+            Path       = $validPath
+            Ensure     = 'Present'
+            Location   = 'LocalMachine'
+            Store      = 'My'
+            Exportable = $True
+            Credential = $testCredential
+            Verbose    = $True
         }
 
         $presentParamsWithFriendlyName = @{
@@ -107,23 +132,24 @@ try
             Ensure       = 'Present'
             Location     = 'LocalMachine'
             Store        = 'My'
-            Verbose      = $true
+            Exportable   = $True
+            Credential   = $testCredential
+            Verbose      = $True
             FriendlyName = $certificateFriendlyName
         }
 
         $absentParams = @{
             Thumbprint = $validThumbprint
-            Path       = $validPath
             Ensure     = 'Absent'
             Location   = 'LocalMachine'
             Store      = 'My'
-            Verbose    = $true
+            Verbose    = $True
         }
 
-        Describe 'DSC_CertificateImport\Get-TargetResource' -Tag 'Get' {
-            Context 'When the certificate exists' {
+        Describe 'DSC_PfxImport\Get-TargetResource' -Tag 'Get' {
+            Context 'When the certificate exists with a private key' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should not throw exception' {
                     {
@@ -139,6 +165,35 @@ try
                     $script:result.Thumbprint | Should -BeExactly $validThumbprint
                     $script:result.Path | Should -BeExactly $validPath
                     $script:result.Ensure | Should -BeExactly 'Present'
+                    $script:result.FriendlyName | Should -BeExactly $certificateFriendlyName
+                }
+
+                It 'Should call the expected mocks' {
+                    Assert-MockCalled `
+                        -CommandName Get-CertificateFromCertificateStore `
+                        -ParameterFilter $getCertificateFromCertificateStore_parameterfilter `
+                        -Exactly -Times 1
+                }
+            }
+
+            Context 'When the certificate exists without private key' {
+                Mock -CommandName Get-CertificateFromCertificateStore `
+                    -MockWith $validCertificateWithoutPrivateKey_mock
+
+                It 'Should not throw exception' {
+                    {
+                        $script:result = Get-TargetResource @presentParams
+                    } | Should -Not -Throw
+                }
+
+                It 'Should return a hashtable' {
+                    $script:result | Should -BeOfType System.Collections.Hashtable
+                }
+
+                It 'Should contain the input values' {
+                    $script:result.Thumbprint | Should -BeExactly $validThumbprint
+                    $script:result.Path | Should -BeExactly $validPath
+                    $script:result.Ensure | Should -BeExactly 'Absent'
                     $script:result.FriendlyName | Should -BeExactly $certificateFriendlyName
                 }
 
@@ -179,7 +234,7 @@ try
             }
         }
 
-        Describe 'DSC_CertificateImport\Test-TargetResource' -Tag 'Test' {
+        Describe 'DSC_PfxImport\Test-TargetResource' -Tag 'Test' {
             Context 'When certificate is not in store but should be' {
                 Mock -CommandName Get-CertificateFromCertificateStore
 
@@ -198,7 +253,7 @@ try
 
             Context 'When certificate is in store and should be' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should return true' {
                     Test-TargetResource @presentParams | Should -Be $true
@@ -207,7 +262,7 @@ try
 
             Context 'When certificate is in store and should be and the FriendlyName is correct' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should return true' {
                     Test-TargetResource @presentParamsWithFriendlyName | Should -Be $true
@@ -225,7 +280,7 @@ try
 
             Context 'When certificate is in store but should not be' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should return false' {
                     Test-TargetResource @absentParams | Should -Be $false
@@ -233,15 +288,15 @@ try
             }
         }
 
-        Describe 'DSC_CertificateImport\Set-TargetResource' -Tag 'Set' {
+        Describe 'DSC_PfxImport\Set-TargetResource' -Tag 'Set' {
             BeforeAll {
                 Mock -CommandName Test-Path -MockWith { $true }
-                Mock -CommandName Import-CertificateEx
+                Mock -CommandName Import-PfxCertificate
                 Mock -CommandName Remove-CertificateFromCertificateStore
                 Mock -CommandName Set-CertificateFriendlyNameInCertificateStore
             }
 
-            Context 'When certificate file exists and certificate should be in the store but is not' {
+            Context 'When PFX file exists and certificate should be in the store but is not' {
                 Mock -CommandName Get-CertificateFromCertificateStore
 
                 It 'Should not throw exception' {
@@ -257,10 +312,10 @@ try
                         -Exactly -Times 1
                 }
 
-                It 'Should call Import-Certificate with expected parameters' {
+                It 'Should call Import-PfxCertificate with expected parameters' {
                     Assert-MockCalled `
-                        -CommandName Import-CertificateEx `
-                        -ParameterFilter $importCertificateEx_parameterfilter `
+                        -CommandName Import-PfxCertificate `
+                        -ParameterFilter $importPfxCertificate_parameterfilter `
                         -Exactly -Times 1
                 }
 
@@ -273,9 +328,9 @@ try
                 }
             }
 
-            Context 'When certificate file exists and certificate should be in the store and is' {
+            Context 'When PFX file exists and certificate should be in the store and is' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should not throw exception' {
                     {
@@ -283,12 +338,12 @@ try
                     } | Should -Not -Throw
                 }
 
-                It 'Should not call Test-Path' {
+                It 'Should not call Test-Path with expected parameters' {
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-Certificate' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate with expected parameters' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
@@ -300,7 +355,7 @@ try
                 }
             }
 
-            Context 'When certificate file exists and certificate should be in the store and is and the friendly name is different' {
+            Context 'When PFX file exists and certificate should be in the store and is and the friendly name is different' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
                     -MockWith $validCertificateWithDifferentFriendlyName_mock
 
@@ -314,8 +369,8 @@ try
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-Certificate with the parameters supplied' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate with the parameters supplied' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should call Set-CertificateFriendlyNameInCertificateStore with expected parameters' {
@@ -330,9 +385,9 @@ try
                 }
             }
 
-            Context 'When certificate file exists and certificate should be in the store and is and the friendly name is the same' {
+            Context 'When PFX file exists and certificate should be in the store and is and the friendly name is the same' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should not throw exception' {
                     {
@@ -344,8 +399,8 @@ try
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-Certificate with the parameters supplied' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate with the parameters supplied' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
@@ -357,9 +412,10 @@ try
                 }
             }
 
-            Context 'When certificate file exists and certificate should not be in the store but is' {
+
+            Context 'When PFX file exists and certificate should not be in the store but is' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
-                    -MockWith $validCertificate_mock
+                    -MockWith $validCertificateWithPrivateKey_mock
 
                 It 'Should not throw exception' {
                     {
@@ -371,8 +427,8 @@ try
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-CertificateEx' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
@@ -387,7 +443,7 @@ try
                 }
             }
 
-            Context 'When certificate file exists and certificate should not be in the store and is not' {
+            Context 'When PFX file exists and certificate should not be in the store and is not' {
                 Mock -CommandName Get-CertificateFromCertificateStore
 
                 It 'Should not throw exception' {
@@ -400,8 +456,8 @@ try
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-CertificateEx' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
@@ -416,7 +472,9 @@ try
                 }
             }
 
-            Context 'When certificate file does not exist and certificate should be in the store' {
+            Context 'When PFX file does not exist and certificate should be in the store' {
+                Mock -CommandName Get-CertificateFromCertificateStore
+
                 Mock -CommandName Test-Path -MockWith { $false }
 
                 It 'Should throw exception' {
@@ -425,15 +483,15 @@ try
                     } | Should -Throw ($script:localizedData.CertificatePfxFileNotFoundError -f $validPath)
                 }
 
-                It 'Should call Test-Path with the parameters supplied' {
+                It 'Should call Test-Path with expected parameters' {
                     Assert-MockCalled `
                         -CommandName Test-Path `
                         -ParameterFilter $testPath_parameterfilter `
                         -Exactly -Times 1
                 }
 
-                It 'Should not call Import-CertificateEx' {
-                    Assert-MockCalled -CommandName Import-CertificateEx -Exactly -Times 0
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {