+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated to use continuous delivery pattern using Azure DevOps - Fixes
   [Issue #215](https://github.com/dsccommunity/CertificateDsc/issues/215).
 - Updated Examples and Module Manifest to be DSC Community from Microsoft.
+- Fix style issues in `Certificate.PDT` and `Certificate.Common` modules.